Feat : 퀘스트 목표 변경 기능을 구현한다 #25

+++ src/main/java/com/ripple/BE/user/dto/QuestDTO.java
@@ -1,9 +1,9 @@
 package com.ripple.BE.user.dto;
 
-public record QuestDTO(Long conceptProgress, Long quizProgress, Long articleProgress) {
+public record QuestDTO(int conceptProgress, int quizProgress, int articleProgress) {
 
     public static QuestDTO toQuestDTO(
-            final Long conceptProgress, final Long quizProgress, final Long articleProgress) {
+            final int conceptProgress, final int quizProgress, final int articleProgress) {
         return new QuestDTO(conceptProgress, quizProgress, articleProgress);
     }
 }
